Criada as classes de CPF, CNPJ, ISBN10, ISBN13
