+++ classe/Fisica.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace classe;
+
 class Fisica extends Usuario{
 
     private $cpf;
